% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization_data_sources.R
\name{add_hyperparameters_parameter_settings}
\alias{add_hyperparameters_parameter_settings}
\title{Add hyperparameters to existing parameter settings}
\usage{
add_hyperparameters_parameter_settings(parameter_setting, lr_sig_hub, gr_hub, ltf_cutoff, algorithm, damping_factor, correct_topology)
}
\arguments{
\item{parameter_setting}{A list of lists. Sublists contains parameters (like data source weights) and novel sublists will be created by this function by adding additional parameters.}

\item{lr_sig_hub}{a number between 0 and 1. 0: no correction for hubiness; 1: maximal correction for hubiness.}

\item{gr_hub}{a number between 0 and 1. 0: no correction for hubiness;  1: maximal correction for hubiness.}

\item{ltf_cutoff}{Ligand-tf scores beneath the "ltf_cutoff" quantile will be set to 0. Default: 0.99 such that only the 1 percent closest tfs will be considered as possible tfs downstream of the ligand of choice.}

\item{algorithm}{Selection of the algorithm to calculate ligand-tf signaling probability scores. Different options: "PPR" (personalized pagerank), "SPL" (shortest path length) and "direct"(just take weights of ligand-signaling network as ligand-tf weights). Default and recommended: PPR.}

\item{damping_factor}{Only relevant when algorithm is PPR. In the PPR algorithm, the damping factor is the probability that the random walker will continue its walk on the graph; 1-damping factor is the probability that the walker will return to the seed node. Default: 0.5.}

\item{correct_topology}{This parameter indicates whether the PPR-constructed ligand-target matrix will be subtracted by a PR-constructed target matrix. TRUE or FALSE.}
}
\value{
A list of lists. Every sub-list contains parameter values for a different parameter.
}
\description{
\code{add_hyperparameters_parameter_settings} will generate a list of lists containing the parameter values that need to be used for model construction.
}
\examples{
\dontrun{
weights_settings_loi = prepare_settings_leave_one_in_characterization(lr_network,sig_network, gr_network, source_weights_df)
weights_settings_loi = lapply(weights_settings_loi,add_hyperparameters_parameter_settings, lr_sig_hub = 0.25,gr_hub = 0.5,ltf_cutoff = 0,algorithm = "PPR",damping_factor = 0.8,correct_topology = TRUE)
}
}
