% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_popularity_bias.R
\name{add_ligand_popularity_measures_to_perfs}
\alias{add_ligand_popularity_measures_to_perfs}
\title{Merge target gene prediction performances with popularity measures of ligands}
\usage{
add_ligand_popularity_measures_to_perfs(performances, ncitations)
}
\arguments{
\item{performances}{A data.frame in which the performance measures for target gene predictions of ligands are denoted}

\item{ncitations}{A data frame denoting the number of times a gene is mentioned in the Pubmed literature. Should at least contain following variables: 'symbol' and 'ncitations'. Default: ncitations (variable contained in this package). See function \code{get_ncitations_genes} for a function that makes this data frame from current Pubmed information.}
}
\value{
A data.frame in which the performance measures for target gene prediction of a ligand are merged with the number of times the ligand is mentioned in the Pubmed literature.
}
\description{
\code{add_ligand_popularity_measures_to_perfs}: Get a data.frame in which the performance measures for target gene prediction of a ligand are merged with the number of times the ligand is mentioned in the Pubmed literature. Serves to investigate popularity bias (i.e. it can be expected that frequenly cited ligands will have better predictive performance because they are better studied).
}
\examples{
\dontrun{
library(dplyr)
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df)
settings = lapply(expression_settings_validation[1:10],convert_expression_settings_evaluation)
ligands = extract_ligands_from_settings(settings)
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands)
performances = bind_rows(lapply(settings,evaluate_target_prediction,ligand_target_matrix))
# ncitations = get_ncitations_genes()
performances_ligand_popularity = add_ligand_popularity_measures_to_perfs(performances,ncitations)
}

}
