% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ligand_to_target.R
\name{add_new_datasource}
\alias{add_new_datasource}
\title{Add a new data source to the model}
\usage{
add_new_datasource(new_source, network, new_weight,source_weights_df)
}
\arguments{
\item{new_source}{A data frame / tibble containing novel interactions of the ligand-receptor, signaling or gene regulatory layer (required columns: from, to, source)}

\item{network}{NULL or a data frame / tibble containing the base network to which you want to add the new data source (required columns: from, to, source)}

\item{new_weight}{a weight value between 0 and 1 to assign to your new data source}

\item{source_weights_df}{A data frame / tibble containing the weights associated to each already included individual data source (required columns: source, weight).}
}
\value{
A list containing 2 elements (network and source_weights_df): the updated network containing your data source and the updated source_weights_df containing the weight of the newly added data source.
}
\description{
\code{add_new_datasource} adds a new data source to one of the ligand-receptor, signaling and gene regulatory data sources.
}
\examples{
 \dontrun{
## Update the lr_network with a new ligand-receptor data source
library(dplyr)
lr_toy = tibble(from = "A", to = "B", source = "toy")
new_lr_network = add_new_datasource(lr_toy, lr_network,1,source_weights_df)
}
}
