% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{alias_to_symbol_seurat}
\alias{alias_to_symbol_seurat}
\title{Convert aliases to official gene symbols in a Seurat Object}
\usage{
alias_to_symbol_seurat(seurat_obj, organism)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{organism}{Is Seurat object data from "mouse" or "human"}
}
\value{
Seurat object
}
\description{
\code{alias_to_symbol_seurat} Convert aliases to official gene symbols in a Seurat Object. Makes use of `convert_alias_to_symbols`
}
\examples{
\dontrun{
seurat_object_lite = readRDS(url("https://zenodo.org/record/3531889/files/seuratObj_test.rds"))
seurat_object_lite = seurat_object_lite \%>\% alias_to_symbol_seurat("human")
}

}
