% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{annotation_data_sources}
\alias{annotation_data_sources}
\title{Annotation table of all data sources used in the NicheNet model}
\format{
A data frame/tibble
\describe{
  \item{source}{Name of the data source}
  \item{database}{Name of the comprehensive database to which the data source belongs}
  \item{type_db}{Type of data source}
  \item{type_interaction}{Interaction type}
  \item{network}{Network layer to which the data source belongs: ligand_receptor, signaling or gene_regulatory}
  }
}
\usage{
annotation_data_sources
}
\description{
A data.frame/tibble describing for each type of data source the "parent" database and type of interactions it countains.
}
\keyword{datasets}
