% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ligand_to_target.R
\name{apply_hub_corrections}
\alias{apply_hub_corrections}
\title{Apply hub corrections to the weighted integrated ligand-signaling and gene regulatory network}
\usage{
apply_hub_corrections(weighted_networks, lr_sig_hub, gr_hub)
}
\arguments{
\item{weighted_networks}{A list of two elements: lr_sig: a data frame/ tibble containg weighted ligand-receptor and signaling interactions (from, to, weight); and gr: a data frame/tibble containng weighted gene regulatory interactions (from, to, weight)}

\item{lr_sig_hub}{a number between 0 and 1. 0: no correction for hubiness; 1: maximal correction for hubiness.}

\item{gr_hub}{a number between 0 and 1. 0: no correction for hubiness;  1: maximal correction for hubiness.}
}
\value{
A list containing 2 elements (lr_sig and gr): the hubiness-corrected integrated weighted ligand-signaling and gene regulatory networks in data frame / tibble format with columns: from, to, weight.
}
\description{
\code{apply_hub_corrections} downweighs the importance of nodes with a lot of incoming links in the ligand-signaling and/or gene regulatory network. Hub correction method according to following equation: \eqn{Wcor =W * D^-h} with \eqn{D} the indegree matrix of the respective network and \eqn{h} the correction factor.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
wn = apply_hub_corrections(weighted_networks, lr_sig_hub= 0.5, gr_hub= 0.5)
}
}
