% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_visualization.R
\name{assign_ligands_to_celltype}
\alias{assign_ligands_to_celltype}
\title{Assign ligands to cell types}
\usage{
assign_ligands_to_celltype(seuratObj, ligands, celltype_col, func.agg=mean, func.assign=function(x) mean(x) + sd(x), slot="data", condition_oi=NULL, condition_col=NULL)
}
\arguments{
\item{seuratObj}{Seurat object}

\item{ligands}{Vector of ligands to assign to cell types}

\item{celltype_col}{Metadata column name in the Seurat object that contains the cell type information}

\item{func.agg}{Function to use to aggregate the expression of a ligand across all cells in a cell type (default = mean)}

\item{func.assign}{Function to use to assign a ligand to a cell type (default = mean + SD)}

\item{condition_oi}{Condition of interest to subset the Seurat object (default = NULL)}

\item{condition_col}{Metadata column name in the Seurat object that contains the condition of interest (default = NULL)}

\item{slot}{Slot in the Seurat object to use (default = "data"). If "data", the normalized counts are first exponentiated before aggregation is performed}
}
\value{
A data frame of two columns, the cell type the ligand has been assigned to (\code{ligand_type}) and the ligand name (\code{ligand})
}
\description{
Assign ligands to a sender cell type, based on the strongest expressing cell type of that ligand. Ligands are only assigned to a cell type if that cell type is the only one to show an expression that is higher than the average + SD. Otherwise, it is assigned to "General".
}
\examples{
\dontrun{
assign_ligands_to_celltype(seuratObj = seuratObj, ligands = best_upstream_ligands[1:20],
                          celltype_col = "celltype", func.agg = mean, func.assign = function(x) {mean(x)+sd(x)},
                          slot = "data", condition_oi = "LCMV", condition_col = "aggregate")
}

}
