% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{bootstrap_ligand_activity_analysis}
\alias{bootstrap_ligand_activity_analysis}
\title{Run ligand activity analysis with bootstrap}
\usage{
bootstrap_ligand_activity_analysis(expressed_genes_receiver, geneset_oi, background_expressed_genes, ligand_target_matrix, potential_ligands, n_iter = 10)
}
\arguments{
\item{expressed_genes_receiver}{Genes expressed in the receiver cell type}

\item{background_expressed_genes}{Character vector of gene symbols of the background, non-affected, genes (can contain the symbols of the affected genes as well).}

\item{ligand_target_matrix}{The NicheNet ligand-target matrix denoting regulatory potential scores between ligands and targets (ligands in columns).}

\item{potential_ligands}{Character vector giving the gene symbols of the potentially active ligands you want to define ligand activities for.}

\item{n_iter}{Number of iterations to perform (Default: 10)}
}
\value{
List of n_iter elements, each element containing the output of predict_ligand_activities for a random gene set
}
\description{
\code{bootstrap_ligand_activity_analysis} Randomly sample a gene set from all expressed genes in the receiver cell type, then perform ligand activity analysis on this gene set. This 'null gene set' has equal length to the gene set of interest.
}
\examples{
\dontrun{
permutations <- bootstrap_ligand_activity_analysis(expressed_genes_receiver, geneset_oi, background_expressed_genes,
                           ligand_target_matrix, potential_ligands, n_iter = 10)
}
}
