% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{calculate_de}
\alias{calculate_de}
\title{Calculate differential expression of one cell type versus all other cell types}
\usage{
calculate_de(seurat_obj, celltype_colname, condition_oi = NA, condition_colname = NA, min.pct = 0.1, assay_oi = "RNA")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{celltype_colname}{Name of the meta data column that indicates the cell type of a cell}

\item{condition_oi}{If provided, subset seurat_obj so DE is only calculated for cells belonging to condition_oi}

\item{condition_colname}{Name of the meta data column that indicates from which group/condition a cell comes from}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations (Default: 0)}

\item{logfc.threshold}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells (Default: 0)}

\item{return.thresh}{Only return markers that have a p-value < return.thresh (Default: 1)}

\item{assay_oi}{Which assay need to be used for DE calculation via `FindMarkers`. Default RNA, alternatives: SCT.}
}
\value{
A dataframe containing the DE results
}
\description{
\code{calculate_de} Calculate differential expression of one cell type versus all other cell types using Seurat::FindAllMarkers. If condition_oi is provided, only consider cells from that condition.
}
\examples{
\dontrun{
seurat_obj = readRDS(url("https://zenodo.org/record/3531889/files/seuratObj.rds"))
seurat_obj$celltype <- make.names(seurat_obj$celltype)
# Calculate cell-type specific markers across conditions
calculate_de(seurat_obj, "celltype")
# Calculate LCMV-specific cell-type markers
calculate_de(seurat_obj, "celltype", condition_oi = "LCMV", condition_colname = "aggregate")
}

}
