% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_nichenet.R
\name{calculate_niche_de_targets}
\alias{calculate_niche_de_targets}
\title{Calculate differential expression of receiver cell type in one niche versus all other niches of interest: focus on finding DE genes}
\usage{
calculate_niche_de_targets(seurat_obj, niches, expression_pct, ltf_cutoff, assay_oi = "SCT")
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{niches}{a list of lists/niches giving the name, senders and receiver celltypes for each nice. Sender and receiver cell types should be part of Idents(seurat_obj).}

\item{expression_pct}{input of `min.pct` of `Seurat::FindMarkers`}

\item{lfc_cutoff}{input of `logfc.threshold` of `Seurat::FindMarkers`}

\item{assay_oi}{Which assay need to be used for DE calculation via `FindMarkers`. Default SCT, alternatives: RNA.}
}
\value{
A tibble containing the DE results of the niches versus each other.
}
\description{
\code{calculate_niche_de_targets} Calculate differential expression of receiver cell type in one niche versus all other niches of interest: focus on finding DE genes
}
\examples{
\dontrun{
seurat_obj = readRDS(url("https://zenodo.org/record/5840787/files/seurat_obj_subset_integrated_zonation.rds"))
niches = list(
"KC_niche" = list(
  "sender" = c("LSECs_portal","Hepatocytes_portal","Stellate cells_portal"),
  "receiver" = c("KCs")),
"MoMac2_niche" = list(
  "sender" = c("Cholangiocytes","Fibroblast 2"),
  "receiver" = c("MoMac2")),
"MoMac1_niche" = list(
  "sender" = c("Capsule fibroblasts","Mesothelial cells"),
 "receiver" = c("MoMac1"))
)
calculate_niche_de_targets(seurat_obj, niches, expression_pct = 0.10, lfc_cutoff = 0.15)
}

}
