% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{calculate_p_value_bootstrap}
\alias{calculate_p_value_bootstrap}
\title{Calculate ligand p-values from the bootstrapped ligand activity analysis}
\usage{
calculate_p_value_bootstrap(bootstrap_results, ligand_activities, metric = "aupr_corrected")
}
\arguments{
\item{bootstrap_results}{Output of \code{\link{bootstrap_ligand_activity_analysis}}}

\item{ligand_activities}{Output of \code{\link{predict_ligand_activities}}}

\item{metric}{Metric to use (Default: "aupr_corrected")}
}
\value{
Data frame with the ligand name, the number of times the bootstrapped value had a higher score than the observed (\code{total}), and the p-value for each ligand, calculated as \code{total/length(bootstrap_results)}
}
\description{
\code{calculate_p_value_bootstrap} Calculate the p-value for each ligand from the bootstrapped ligand activity analysis
}
\examples{
\dontrun{
permutations <- bootstrap_ligand_activity_analysis(expressed_genes_receiver, geneset_oi, background_expressed_genes,
                           ligand_target_matrix, potential_ligands, n_iter = 10)
p_values <- calculate_p_value_bootstrap(permutations, ligand_activities, metric = "aupr_corrected")
}
}
