% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{classification_evaluation_continuous_pred_wrapper}
\alias{classification_evaluation_continuous_pred_wrapper}
\title{Assess how well classification predictions accord to the expected response}
\usage{
classification_evaluation_continuous_pred_wrapper(response_prediction_tibble)
}
\arguments{
\item{response_prediction_tibble}{Tibble with columns "response" and "prediction" (e.g. output of function `assess_rf_class_probabilities`)}
}
\value{
A tibble showing several classification evaluation metrics.
}
\description{
\code{classification_evaluation_continuous_pred_wrapper} Assess how well classification predictions accord to the expected response.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
ligands = list("TNF","BMP2","IL4")
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands, ltf_cutoff = 0, algorithm = "PPR", damping_factor = 0.5, secondary_targets = FALSE)
potential_ligands = c("TNF","BMP2","IL4")
geneset = c("SOCS2","SOCS3", "IRF1")
background_expressed_genes = c("SOCS2","SOCS3","IRF1","ICAM1","ID1","ID2","ID3")
fold1_rf_prob = assess_rf_class_probabilities(round = 1,folds = 2,geneset = geneset,background_expressed_genes = background_expressed_genes ,ligands_oi = potential_ligands,ligand_target_matrix = ligand_target_matrix)
}

}
