% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ligand_to_target.R
\name{construct_ligand_target_matrix}
\alias{construct_ligand_target_matrix}
\title{Construct ligand-target matrix}
\usage{
construct_ligand_target_matrix(lr_network, sig_network, gr_network)
}
\arguments{
\item{lr_network}{A list.}

\item{sig_network}{A list.}

\item{gr_network}{A list.}
}
\value{
A matrix containing the probability scores of ligand-target links.
}
\description{
\code{construct_ligand_target_matrix} construct a ligand-target matrix from input (weighted) networks.
}
\examples{
## Generate the ligand-target matrix from loaded weighted_networks
construct_ligand_target_matrix(lr_network, sig_network, gr_network)

}
