% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ligand_to_target.R
\name{construct_tf_target_matrix}
\alias{construct_tf_target_matrix}
\title{Construct a tf-target matrix.}
\usage{
construct_tf_target_matrix(weighted_networks, tfs_as_cols = FALSE, standalone_output = FALSE)
}
\arguments{
\item{weighted_networks}{A list of two elements: lr_sig: a data frame/ tibble containg weighted ligand-receptor and signaling interactions (from, to, weight); and gr: a data frame/tibble containng weighted gene regulatory interactions (from, to, weight)}

\item{tfs_as_cols}{Indicate whether ligands should be in columns of the matrix and target genes in rows or vice versa. Default: FALSE}

\item{standalone_output}{Indicate whether the ligand-tf matrix should be formatted in a way convenient to use alone (with gene symbols as row/colnames). Default: FALSE}
}
\value{
A matrix containing tf-target regulatory weights.
}
\description{
\code{construct_tf_target_matrix} Convert integrated gene regulatory weighted network into matrix format.
}
\examples{
 \dontrun{
## Generate the ligand-target matrix from loaded weighted_networks
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
tf_target = construct_tf_target_matrix(weighted_networks, tfs_as_cols = TRUE, standalone_output = TRUE)
}
}
