% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_ligand_to_target.R
\name{construct_weighted_networks}
\alias{construct_weighted_networks}
\title{Construct weighted layer-specific networks}
\usage{
construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
}
\arguments{
\item{lr_network}{A data frame / tibble containing ligand-receptor interactions (required columns: from, to, source)}

\item{sig_network}{A data frame / tibble containing signaling interactions (required columns: from, to, source)}

\item{gr_network}{A data frame / tibble containing gene regulatory interactions (required columns: from, to, source)}

\item{source_weights_df}{A data frame / tibble containing the weights associated to each individual data source. Sources with higher weights will contribute more to the final model performance (required columns: source, weight). Note that only interactions described by sources included here, will be retained during model construction.}
}
\value{
A list containing 3 elements (lr, sig, gr): the integrated weighted ligand-receptor, signaling and gene regulatory networks in data frame / tibble format.
}
\description{
\code{construct_weighted_networks} construct layer-specific weighted integrated networks from input source networks via weighted aggregation.
}
\examples{
## Generate the weighted networks from input source networks
#' weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)


}
