% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_target_prediction.R
\name{convert_expression_settings_evaluation_regression}
\alias{convert_expression_settings_evaluation_regression}
\title{Convert expression settings to correct settings format for evaluation of target gene log fold change prediction (regression).}
\usage{
convert_expression_settings_evaluation_regression(setting)
}
\arguments{
\item{setting}{A list containing the following elements: .$name: name of the setting; .$from: name(s) of the ligand(s) active in the setting of interest; .$diffexp: data frame or tibble containing at least 3 variables= $gene, $lfc (log fold change treated vs untreated) and $qval (fdr-corrected p-value)}
}
\value{
A list with following elements: $name, $from, $response. $response will be a gene-named numeric vector of log fold change values.
}
\description{
\code{convert_expression_settings_evaluation_regression} Converts expression settings to correct settings format for evaluation of target gene log fold change prediction (regression).
}
\examples{
\dontrun{
settings = lapply(expression_settings_validation,convert_expression_settings_evaluation_regression)
}
}
