% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_target_prediction.R
\name{convert_gene_list_settings_evaluation}
\alias{convert_gene_list_settings_evaluation}
\title{Convert gene list to correct settings format for evaluation of target gene prediction.}
\usage{
convert_gene_list_settings_evaluation(gene_list, name, ligands_oi, background)
}
\arguments{
\item{gene_list}{A character vector of target gene names}

\item{name}{The name that will be given to the setting}

\item{ligands_oi}{The possibly active ligands}

\item{background}{A character vector of names of genes that are not target genes. If genes present in the gene list are in this vector of background gene names, these genes will be removed from the background.}
}
\value{
A list with following elements: $name, $from, $response
}
\description{
\code{convert_gene_list_settings_evaluation} Converts a gene list to correct settings format for evaluation of target gene prediction.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df)
all_genes = unique(c(weighted_networks$gr$from,weighted_networks$gr$to,weighted_networks$lr_sig$from, weighted_networks$lr_sig$to))
gene_list = c("ID1","ID2","ID3")
setting = list(convert_gene_list_settings_evaluation(gene_list = c("ID1","ID2","ID3"), name = "test",ligands_oi = "TGFB1", background = all_genes))
}
}
