% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{convert_mouse_to_human_symbols}
\alias{convert_mouse_to_human_symbols}
\title{Convert mouse gene symbols to their human one-to-one orthologs.}
\usage{
convert_mouse_to_human_symbols(symbols, version = 1)
}
\arguments{
\item{symbols}{A character vector of official mouse gene symbols}

\item{version}{Indicates which version of the mouse-human annotations should be used: Default 1. In April 2022, and nichenetr 2.0., the default will change to 2.}
}
\value{
A character vector of official human gene symbols (one-to-one orthologs of the input mouse gene symbols).
}
\description{
\code{convert_mouse_to_human_symbols} Convert mouse gene symbols to their human one-to-one orthologs.
}
\examples{
library(dplyr)
mouse_symbols = c("Tnf","Ifng")
human_symbols = mouse_symbols \%>\% convert_mouse_to_human_symbols()
}
