% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{convert_single_cell_expression_to_settings}
\alias{convert_single_cell_expression_to_settings}
\title{Prepare single-cell expression data to perform ligand activity analysis}
\usage{
convert_single_cell_expression_to_settings(cell_id, expression_matrix, setting_name, setting_from, regression = FALSE)
}
\arguments{
\item{cell_id}{Identity of the cell of interest}

\item{expression_matrix}{Gene expression matrix of single-cells}

\item{setting_name}{Name of the dataset}

\item{setting_from}{Character vector giving the gene symbols of the potentially active ligands you want to define ligand activities for.}

\item{regression}{Perform regression-based ligand activity analysis (TRUE) or classification-based ligand activity analysis (FALSE) by considering the genes expressed higher than the 0.975 quantiles as genes of interest. Default: FALSE.}
}
\value{
A list with slots $name, $from and $response respectively containing the setting name, potentially active ligands and the response to predict (whether genes belong to gene set of interest; i.e. most strongly expressed genes in a cell)
}
\description{
\code{convert_single_cell_expression_to_settings} Prepare single-cell expression data to perform ligand activity analysis
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
ligands = list("TNF","BMP2","IL4")
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands, ltf_cutoff = 0, algorithm = "PPR", damping_factor = 0.5, secondary_targets = FALSE)
potential_ligands = c("TNF","BMP2","IL4")
genes = c("SOCS2","SOCS3","IRF1","ICAM1","ID1","ID2","ID3")
cell_ids = c("cell1","cell2")
expression_scaled = matrix(rnorm(length(genes)*2, sd = 0.5, mean = 0.5), nrow = 2)
rownames(expression_scaled) = cell_ids
colnames(expression_scaled) = genes
settings = convert_single_cell_expression_to_settings(cell_id = cell_ids[1], expression_matrix = expression_scaled, setting_name = "test", setting_from = potential_ligands)
}

}
