% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_visualization.R
\name{draw_circos_plot}
\alias{draw_circos_plot}
\title{Draw a circos plot}
\usage{
draw_circos_plot(vis_circos_obj, transparency = FALSE, args.circos.text = list(), ...)
}
\arguments{
\item{vis_circos_obj}{Object returned by \code{\link{prepare_circos_visualization}}}

\item{transparency}{Logical indicating whether the transparency of the links will correspond to the ligand-target potential score (default: FALSE)}

\item{args.circos.text}{List of arguments to pass to \code{\link{circos.text}} (by default, the text size is set to 1)}

\item{...}{Additional arguments to pass to \code{\link{chordDiagram}}}
}
\value{
A circos plot
}
\description{
Draw a circos plot
}
\examples{
\dontrun{
# Default
draw_circos_plot(vis_circos_obj, transparency = FALSE)

# Transparency
draw_circos_plot(vis_circos_obj, transparency = TRUE)

# Make text smaller
draw_circos_plot(vis_circos_obj, transparency = TRUE, args.circos.text = list(cex = 0.5))

# Don't sort links of each ligand based on widths (not recommended)
draw_circos_plot(vis_circos_obj, transparency = TRUE, args.circos.text = list(cex = 0.5), link.sort = FALSE)
}
}
