% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_ligand_prediction}
\alias{evaluate_ligand_prediction}
\title{Construct ligand-target matrix}
\usage{
evaluate_ligand_prediction(lr_network, sig_network, gr_network)
}
\arguments{
\item{lr_network}{A list.}

\item{sig_network}{A list.}

\item{gr_network}{A list.}
}
\value{
A matrix containing the probability scores of ligand-target links.
}
\description{
\code{evaluate_ligand_prediction} construct a ligand-target matrix from input (weighted) networks.
}
\examples{
## Generate the ligand-target matrix from loaded weighted_networks
evaluate_ligand_prediction(mtcars, mtcars, mtcars)

}
