% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Exprs_lsec}
\alias{Exprs_lsec}
\title{ExpressionSet object containing gene expression data from liver sinusoidal endothelial cells (LSECs) in wild type situation and also 12 hours and 36 hours after Kupffer cells were killed by giving Diphteria toxin to KC-DTR mice.}
\format{An object of class \code{ExpressionSet} with 12173 rows and 12 columns.}
\usage{
Exprs_lsec
}
\description{
Formal class 'ExpressionSet'
}
\keyword{datasets}
