% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Exprs_mono_kc}
\alias{Exprs_mono_kc}
\title{ExpressionSet object containing gene expression data from bone marrow monocytes (data from 'van de Laar et al, 2016') and liver monocytes that differentiate into Kupffer cells after Kupffer cells were killed by giving Diphteria toxin to KC-DTR mice. Time points in differentiation process after treatment with toxin: 1 day, 3 days, 7 days, 15 days and 30 days.}
\format{An object of class \code{ExpressionSet} with 20908 rows and 18 columns.}
\usage{
Exprs_mono_kc
}
\description{
Formal class 'ExpressionSet'
}
\keyword{datasets}
