% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_ligand_to_target.R
\name{extract_top_fraction_targets}
\alias{extract_top_fraction_targets}
\title{Get the predicted top n percentage target genes of a ligand of interest}
\usage{
extract_top_fraction_targets(ligand_oi,top_fraction,ligand_target_matrix,ligands_position = "cols")
}
\arguments{
\item{ligand_oi}{The ligand of interest of which top target genes should be returned}

\item{top_fraction}{A number between 0 and 1 indicating which top fraction of target genes should be returned.}

\item{ligand_target_matrix}{A matrix of ligand-target probabilty scores.}

\item{ligands_position}{Indicate whether the ligands in the ligand-target matrix are in the rows ("rows") or columns ("cols"). Default: "cols"}
}
\value{
A named numeric vector of ligand-target gene probability scores of the top target genes.
}
\description{
\code{extract_top_fraction_targets} Get the predicted top n percentage target genes of a ligand of interest.
}
\examples{
\dontrun{
## Generate the ligand-target matrix from loaded weighted_networks
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
ligands = list("TNF","BMP2",c("IL4","IL13"))
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands)
targets = extract_top_fraction_targets("BMP2",0.01,ligand_target_matrix)
}
}
