% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneinfo_alias_human}
\alias{geneinfo_alias_human}
\title{Gene annotation information: version 2 - january 2022 - suited for alias conversion}
\format{
A data frame/tibble
\describe{
  \item{symbol}{human gene symbol}
  \item{entrez}{human gene entrez}
  \item{alias}{human gene alias}
  }
}
\usage{
geneinfo_alias_human
}
\description{
A data.frame/tibble describing HGNC human gene symbols, their entrez ids and potential aliases.
}
\keyword{datasets}
