% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneinfo_alias_mouse}
\alias{geneinfo_alias_mouse}
\title{Gene annotation information: version 2 - january 2022 - suited for alias conversion}
\format{
A data frame/tibble
\describe{
  \item{symbol}{mouse gene symbol}
  \item{entrez}{mouse gene entrez}
  \item{alias}{mouse gene alias}
  }
}
\usage{
geneinfo_alias_mouse
}
\description{
A data.frame/tibble describing MGI mouse gene symbols, their entrez ids and potential aliases.
}
\keyword{datasets}
