% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneinfo_human}
\alias{geneinfo_human}
\title{Gene annotation information}
\format{A data frame/tibble
\describe{
  \item{symbol}{human gene symbol}
  \item{entrez}{human gene entrez}
  \item{entrez_mouse}{mouse homolog gene entrez}
  \item{symbol_mouse}{mouse homolog gene symbol}
  }}
\usage{
geneinfo_human
}
\description{
A data.frame/tibble describing weight associated to each data source. These weights will be used for weighted aggregagion of source networks
}
\keyword{datasets}
