% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{generate_info_tables}
\alias{generate_info_tables}
\title{Generate tables used for \code{generate_prioritization_tables}}
\usage{
generate_info_tables(
  seuratObj,
  celltype_colname,
  senders_oi,
  receivers_oi,
  lr_network_filtered,
  condition_colname = NULL,
  condition_oi = NULL,
  condition_reference = NULL,
  scenario = "case_control",
  assay_oi = NULL,
  ...
)
}
\arguments{
\item{seuratObj}{Seurat object}

\item{celltype_colname}{Name of the meta data column that indicates the cell type of a cell}

\item{senders_oi}{Default NULL: all celltypes will be considered as senders. If you want to select specific senders of interest: you can add this here as character vector.}

\item{receivers_oi}{Default NULL: all celltypes will be considered as receivers If you want to select specific receivers of interest: you can add this here as character vector.}

\item{lr_network_filtered}{Ligand-receptor network that has been filtered to only contain ligands and receptors that are expressed}

\item{condition_colname}{Name of the meta data column that indicates from which group/condition a cell comes from}

\item{condition_oi}{If provided, subset seurat_obj so DE is only calculated for cells belonging to condition_oi}

\item{condition_reference}{Reference condition for condition specificity calculation}

\item{scenario}{"case_control" or "one_condition": if "case_control", calculate condition specificity. If "one_condition", only calculate cell type specificity.}

\item{assay_oi}{Which assay need to be used for DE calculation. If NULL, will use \code{DefaultAssay}}

\item{...}{Arguments passed to \code{FindAllMarkers}, \code{FindMarkers}, and \code{AverageExpression}}
}
\value{
List of dataframes containing sender-receiver DE, sender-receiver expression, and condition DE
}
\description{
Calculate differential expression, average expression, and condition specificity of ligands and receptors.
}
