% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_network_extraction.R
\name{get_active_ligand_receptor_network}
\alias{get_active_ligand_receptor_network}
\title{Get active ligand-receptor network for cellular interaction between a sender and receiver cell.}
\usage{
get_active_ligand_receptor_network(expression_sender, expression_receiver, lr_network, expression_cutoff_sender = 0, expression_cutoff_receiver = 0)
}
\arguments{
\item{expression_sender}{A named numeric vector of gene expression levels (absolute or logfc) for the signaling cell that sends extracellular signals to the receiver cell}

\item{expression_receiver}{A named numeric vector of gene expression levels (absolute or logfc) for the receiver cell that receives extracellular signals from the sender cell}

\item{lr_network}{A data frame / tibble containing ligand-receptor interactions (required columns: from, to). Can be both unweighted and weighted.}

\item{expression_cutoff_sender}{The cutoff on expression value for the sender cell: ligands will be considered active if their expression is higher than the cutoff. Default: 0.}

\item{expression_cutoff_receiver}{The cutoff on expression value for the receiver cell: receptors will be considered active if their expression is higher than the cutoff. Default: 0.}
}
\value{
A data frame containing at least the variables from, to, sender_expression, receiver_expression. In this network, the active ligand-receptor interactions in the system of interest are shown.
}
\description{
\code{get_active_ligand_receptor_network} Get active ligand-receptor network by looking for which ligands are expressed in a sender/signaling cell and which receptors are expressed in the receiver cell. Instead of looking at absolute expression, it is possible as well to extract a ligand-receptor network of differentially expressed ligands and receptors if a vector of log2 fold change values is used as input.
}
\examples{
library(Biobase)
library(dplyr)
mousesymbol2humansymbol = mapper(geneinfo_human,"symbol","symbol_mouse")
expression_vector_sender = Exprs_lsec[,Exprs_lsec$celltype == "LSEC_12h"] \%>\% apply(1,mean)
expression_vector_receiver = Exprs_mono_kc[,Exprs_mono_kc$celltype == "BM_mono"] \%>\% apply(1,mean)
names(expression_vector_sender) = names(expression_vector_sender) \%>\% mousesymbol2humansymbol[.]
names(expression_vector_receiver) = names(expression_vector_receiver) \%>\% mousesymbol2humansymbol[.]
expression_vector_sender = expression_vector_sender \%>\% .[!is.na(names(.))]
expression_vector_receiver = expression_vector_receiver \%>\% .[!is.na(names(.))]
weighted_lr_network = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df, n_output_networks = 3) \%>\% .$lr
lsec_mono_lr_network = get_active_ligand_receptor_network(expression_vector_sender,expression_vector_receiver,weighted_lr_network,expression_cutoff_sender = 0, expression_cutoff_receiver = 4)
}
