% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_network_extraction.R
\name{get_active_signaling_network}
\alias{get_active_signaling_network}
\title{Get active signaling network in a receiver cell.}
\usage{
get_active_signaling_network(expression_receiver, sig_network, expression_cutoff_receiver = 0)
}
\arguments{
\item{expression_receiver}{A named numeric vector of gene expression levels (absolute or logfc) for the receiver cell that receives extracellular signals from the sender cell}

\item{sig_network}{A data frame / tibble containing signaling interactions (required columns: from, to). Can be both unweighted and weighted.}

\item{expression_cutoff_receiver}{The cutoff on expression value for the receiver cell: receptors will be considered active if their expression is higher than the cutoff. Default: 0.}
}
\value{
A data frame containing at least the variables from, to, receiver_expression. In this network, the active signaling interactions in the system of interest are shown.
}
\description{
\code{get_active_signaling_network} Get active signaling network by looking for which signaling-related genes are expressed in the receiver cell. Instead of looking at absolute expression, it is possible as well to extract a signaling network of differentially expressed signaling mediators if a vector of log2 fold change values is used as input.
}
\examples{
\dontrun{
library(Biobase)
library(dplyr)
mapper = function(df, value_col, name_col) setNames(df[[value_col]], df[[name_col]])
mousesymbol2humansymbol = mapper(geneinfo_human,"symbol","symbol_mouse")
expression_vector_receiver = Exprs_mono_kc[,Exprs_mono_kc$celltype == "BM_mono"] \%>\% apply(1,mean)
names(expression_vector_receiver) = names(expression_vector_receiver) \%>\% mousesymbol2humansymbol[.]
expression_vector_receiver = expression_vector_receiver \%>\% .[!is.na(names(.))]
mono_sig_network = get_active_signaling_network(expression_vector_receiver,sig_network,expression_cutoff_receiver = 4)
}
}
