% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{get_all_networks}
\alias{get_all_networks}
\title{Get the complete ligand-receptor, signaling and gene regulatory network.}
\usage{
get_all_networks()
}
\value{
a list containing the following elements: $lr: The true complete ligand-receptor network, $sig: the true complete signaling network, $gr: the true complete gene regulatory network.
}
\description{
\code{get_all_networks} will return a list containing the complete ligand-receptor network, signaling network and gene regulatory network.
}
\examples{
\dontrun{
all_networks = get_all_networks()
}
}
