% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{get_expressed_genes}
\alias{get_expressed_genes}
\alias{get_expressed_genes.default}
\alias{get_expressed_genes.Seurat}
\title{Determine expressed genes of a cell type from an input object}
\usage{
get_expressed_genes(object, pct = 0.1, ...)

\method{get_expressed_genes}{default}(object, celltype_annot, celltype_oi, pct = 0.1)

\method{get_expressed_genes}{Seurat}(seurat_obj, ident, assay_oi = NULL, pct = 0.1, ...)
}
\arguments{
\item{object}{Input matrix with rows as genes and columns as cells}

\item{pct}{We consider genes expressed if they are expressed in at least a specific fraction of cells of the given cluster(s). This number indicates this fraction. Default: 0.10. Choice of this parameter is important and depends largely on the used sequencing platform. We recommend to require a lower fraction (like the default 0.10) for 10X data than for e.g. Smart-seq2 data.}

\item{...}{additional parameters passed to \code{GetAssayData} (in case the slot/layer needs to be specified)}

\item{celltype_annot}{Vector of cell type annotations}

\item{celltype_oi}{Cell type(s) to be considered}

\item{seurat_obj}{Single-cell expression or spatial dataset as Seurat object}

\item{ident}{Name of cluster identity/identities of cells}

\item{assay_oi}{If wanted: specify yourself which assay to look for. If not (NULL), the \code{DefaultAssay} of the Seurat object is used.}
}
\value{
A character vector with the gene symbols of the expressed genes
}
\description{
Return the genes that are expressed in given cell cluster(s) based on the fraction of cells in the cluster(s) that should express the cell.
}
\examples{
\dontrun{
get_expressed_genes(seurat_obj = seuratObj, ident = "CD8 T", pct = 0.10)
}

}
