% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_data.R
\name{get_expression_settings_validation}
\alias{get_expression_settings_validation}
\title{Get the complete set of ligand-treatment expression datasets.}
\usage{
get_expression_settings_validation()
}
\value{
The true complete set of ligand-treatment expression datasets.
}
\description{
\code{get_expression_settings_validation} will return the complete set of ligand-treatment expression datasets.
}
\examples{
\dontrun{
expression_settings_validation = get_expression_settings_validation()
}
}
