% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_visualization.R
\name{get_ligand_target_links_oi}
\alias{get_ligand_target_links_oi}
\title{Get ligand-target links of interest}
\usage{
get_ligand_target_links_oi(ligand_type_indication_df, active_ligand_target_links_df, cutoff = 0.40)
}
\arguments{
\item{ligand_type_indication_df}{Dataframe with column names \code{ligand_type}) and \code{ligand}, from the function \code{\link{assign_ligands_to_celltype}}}

\item{active_ligand_target_links_df}{Dataframe with weighted ligand-target links from the function \code{\link{get_ligand_target_links}}, and an additional column \code{target_type} that indicates the grouping of target genes}

\item{cutoff}{Quantile to filter ligand-target links (default = 0.40, meaning 40\% of the lowest weighted ligand-target links are removed)}
}
\value{
A dataframe with ligand-target links with weights above a certain cutoff
}
\description{
Filter ligand-target links based on a cutoff
}
\examples{
\dontrun{
active_ligand_target_links_df <- lapply(best_upstream_ligands, get_weighted_ligand_target_links,
                                         geneset = geneset_oi,
                                         ligand_target_matrix = ligand_target_matrix,
                                         n = 200)
active_ligand_target_links_df <- drop_na(bind_rows(active_ligand_target_links_df))
ligand_type_indication_df <- assign_ligands_to_celltype(seuratObj = seuratObj, ligands = best_upstream_ligands[1:20])
circos_links <- get_ligand_target_links_oi(ligand_type_indication_df,
                                           active_ligand_target_links_df \%>\% mutate(target_type = "LCMV-DE"),
                                           cutoff = 0.40)
}

}
