% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_nichenet.R
\name{get_prioritization_tables}
\alias{get_prioritization_tables}
\title{Use the information from the niche- and spatial differential expression analysis of ligand-senders and receptor-receivers pairs, in addition to the ligand activity prediction and ligand-target inferernce, in order to make a final ligand-receptor and ligand-target prioritization table.}
\usage{
get_prioritization_tables(output_nichenet_analysis, prioritizing_weights)
}
\arguments{
\item{output_nichenet_analysis}{List containing following data frames: DE_sender_receiver, ligand_scaled_receptor_expression_fraction_df, sender_spatial_DE_processed,
receiver_spatial_DE_processed, ligand_activities_targets, DE_receiver_processed_targets, exprs_tbl_ligand, exprs_tbl_receptor, exprs_tbl_target}

\item{prioritizing_weights}{Named numeric vector in the form of:
#' prioritizing_weights = c("scaled_ligand_score" = 5,}
}
\value{
A list containing a prioritization table for ligand-receptor interactions, and one for ligand-target interactions
}
\description{
\code{get_prioritization_tables} Use the information from the niche- and spatial differential expression analysis of ligand-senders and receptor-receivers pairs, in addition to the ligand activity prediction and ligand-target inferernce, in order to make a final ligand-receptor and ligand-target prioritization table.
}
\examples{
\dontrun{
prioritizing_weights = c("scaled_ligand_score" = 5,
output_nichenet_analysis = list(DE_sender_receiver = DE_sender_receiver, ligand_scaled_receptor_expression_fraction_df = ligand_scaled_receptor_expression_fraction_df, sender_spatial_DE_processed = sender_spatial_DE_processed, receiver_spatial_DE_processed = receiver_spatial_DE_processed,
prioritization_tables = get_prioritization_tables(output_nichenet_analysis, prioritizing_weights)
}

}
