% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_popularity_bias.R
\name{get_slope_ligand_popularity}
\alias{get_slope_ligand_popularity}
\title{Regression analysis between ligand popularity and target gene predictive performance}
\usage{
get_slope_ligand_popularity(metric, performances)
}
\arguments{
\item{metric}{The name of the performance metric of which the trend with the popularity of the ligand should be calculated.}

\item{performances}{A data.frame in which the performance measures for target gene predictions of ligands are denoted together with the popularity of the ligand. (should contain at least following variables: ligand, ncitations and the metric of interest)}
}
\value{
A data.frame in which the regression coefficient estimate, p-value and corresponding R-squared value are shown for the regression analysis to investigate the trend between a particular classficiation evaluation metric and the popularity of the ligand.
}
\description{
\code{get_slope_ligand_popularity}: Performs regression analysis to investigate the trend between a particular classficiation evaluation metric and the popularity of the ligand.
}
\examples{
\dontrun{
library(dplyr)
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df)
settings = lapply(expression_settings_validation[1:10],convert_expression_settings_evaluation)
ligands = extract_ligands_from_settings(settings)
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands)
performances = bind_rows(lapply(settings,evaluate_target_prediction,ligand_target_matrix))
# ncitations = get_ncitations_genes()
performances_ligand_popularity = add_ligand_popularity_measures_to_perfs(performances,ncitations)
slopes_auroc = get_slope_ligand_popularity("auroc",performances_ligand_popularity)
slopes_df = performances_ligand_popularity \%>\% select(-setting,-ligand,-ncitations) \%>\% colnames() \%>\% lapply(.,get_slope_ligand_popularity,performances_ligand_popularity) \%>\% bind_rows()
}


}
