% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_popularity_bias.R
\name{get_slope_target_gene_popularity_ligand_prediction}
\alias{get_slope_target_gene_popularity_ligand_prediction}
\title{Regression analysis between target gene popularity and ligand activity predictive performance}
\usage{
get_slope_target_gene_popularity_ligand_prediction(metric,performances)
}
\arguments{
\item{metric}{The name of the performance metric of which the trend with the popularity of the target genes should be calculated.}

\item{performances}{A data.frame in which the performance measures for ligand activity predictions of ligands are denoted together with the popularity bin of the target genes for which predictions were evaluated (should contain at least following variables: target_bin_id and the metric of interest)}
}
\value{
A data.frame in which the regression coefficient estimate, p-value and corresponding R-squared value are shown for the regression analysis to investigate the trend between a particular classficiation evaluation metric and the popularity of the target genes.
}
\description{
\code{get_slope_target_gene_popularity_ligand_prediction}: Performs regression analysis to investigate the trend between a particular classficiation evaluation metric (ligand activity prediction) and the popularity of target genes.
}
\examples{
\dontrun{
library(dplyr)
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df)
settings = lapply(expression_settings_validation[1:10],convert_expression_settings_evaluation)
ligands = extract_ligands_from_settings(settings)
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands)
# ncitations = get_ncitations_genes()
performances_target_bins_popularity = evaluate_ligand_prediction_per_bin(5,settings,ligand_target_matrix,ncitations)
slopes_auroc = get_slope_target_gene_popularity_ligand_prediction("auroc",performances_target_bins_popularity)
slopes_df = performances_target_bins_popularity \%>\% select(-importance_measure, -target_bin_id) \%>\% colnames() \%>\% lapply(.,get_slope_target_gene_popularity_ligand_prediction,performances_target_bins_popularity) \%>\% bind_rows()
}

}
