% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{get_weighted_ligand_receptor_links}
\alias{get_weighted_ligand_receptor_links}
\title{Get the weighted ligand-receptor links between a possible ligand and its receptors}
\usage{
get_weighted_ligand_receptor_links(
  best_upstream_ligands,
  expressed_receptors,
  lr_network,
  weighted_networks_lr_sig
)
}
\arguments{
\item{best_upstream_ligands}{Character vector containing ligands of interest.}

\item{expressed_receptors}{Character vector of receptors expressed in the cell type of interest.}

\item{lr_network}{A data frame with two columns, \code{from} and \code{to}, containing the ligand-receptor interactions.}

\item{weighted_networks_lr_sig}{A data frame with three columns, \code{from}, \code{to} and \code{weight}, containing the ligand-receptor interactions and their weights.}
}
\description{
\code{get_weighted_ligand_receptor_links} Get the weighted ligand-receptor links between a possible ligand and its receptors
}
\examples{
\dontrun{
  ligand_receptor_links_df <- get_weighted_ligand_receptor_links(best_upstream_ligands, expressed_receptors, lr_network, weighted_networks$lr_sig)
}

}
