% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gr_network}
\alias{gr_network}
\title{Gene regulatory network}
\format{
A data frame/tibble
\describe{
  \item{from}{regulatory gene}
  \item{to}{target gene}
  \item{source}{name of data source describing the interaction}
  \item{database}{name of the more comprehensive database the data source is part of}
  }
}
\usage{
gr_network
}
\description{
A data.frame/tibble describing gene regulatory interactions such as TF-target interactions
}
\keyword{datasets}
