% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Say hello to whoever you want!}
\usage{
hello(a = "world")
}
\arguments{
\item{a}{Character vector. The name of whoever you want to greet. must follow following format
\describe{
 \item{grn}{Gene regulatory network}
 \item{sig}{signaling network}
 }}
}
\value{
a string
}
\description{
return hello to everyone, everything you want
}
\details{
blababjlballdfdfjdlfjdfkdfjdmfkldjfkdqjfmsldqjfmkldqfjsdqmlfjsdqmlfjsdqmfjqmfjsdqmfjqmj
dfdfdfdfddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddddfdfdf
}
\examples{
a = hello("Hadley")
print(a)
hello("World")
}
