% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{infer_ligand_target_links}
\alias{infer_ligand_target_links}
\title{Infer active ligand target links between possible lignands and target genes of interest}
\usage{
infer_ligand_target_links(ligands_oi, targets_oi, background_expressed_genes, ligand_target_matrix, cutoff = 0.025)
}
\arguments{
\item{ligands_oi}{Character vector giving the gene symbols of the potentially active ligands you want to define ligand activities for.}

\item{targets_oi}{Character vector of the gene symbols of genes of which the expression is potentially affected by ligands from the interacting cell.}

\item{background_expressed_genes}{Character vector of gene symbols of the background, non-affected, genes (can contain the symbols of the affected genes as well).}

\item{ligand_target_matrix}{The NicheNet ligand-target matrix denoting regulatory potential scores between ligands and targets (ligands in columns).}

\item{cutoff}{Genes with a regulatory potential score for a ligand not part of the top percentage scores, will not be considered as targets. Default: 0.025.}
}
\value{
A matrix giving the ligand-target regulatory potential scores between ligands of interest and their targets genes part of the gene set of interest.
}
\description{
\code{infer_ligand_target_links} Infer active ligand target links between possible lignands and genes belonging to a gene set of interest.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
ligands = list("TNF","BMP2","IL4")
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands, ltf_cutoff = 0, algorithm = "PPR", damping_factor = 0.5, secondary_targets = FALSE)
potential_ligands = c("TNF","BMP2","IL4")
geneset = c("SOCS2","SOCS3", "IRF1")
background_expressed_genes = c("SOCS2","SOCS3","IRF1","ICAM1","ID1","ID2","ID3")
active_ligand_target_links = infer_ligand_target_links(ligands_oi = potential_ligands, targets_oi = geneset, background_expressed_genes = background_expressed_genes, ligand_target_matrix = ligand_target_matrix)
}

}
