% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_nichenet_plotting.R
\name{make_ligand_activity_target_exprs_plot}
\alias{make_ligand_activity_target_exprs_plot}
\title{make_ligand_activity_target_exprs_plot}
\usage{
make_ligand_activity_target_exprs_plot(receiver_oi, prioritized_tbl_oi, prioritization_tbl_ligand_receptor, prioritization_tbl_ligand_target, exprs_tbl_ligand, exprs_tbl_target, lfc_cutoff, plot_legend = TRUE, heights = NULL, widths = NULL)
}
\arguments{
\item{receiver_oi}{Name of the receiver cell type of interest}

\item{prioritized_tbl_oi}{Dataframe with the ligand-receptor interactions that should be visualized}

\item{prioritization_tbl_ligand_receptor}{$prioritization_tbl_ligand_receptor slot of `get_prioritization_tables`}

\item{prioritization_tbl_ligand_target}{$prioritization_tbl_ligand_target slot of `get_prioritization_tables`}

\item{exprs_tbl_ligand}{Dataframe with the expression values for the ligands in the sender cell types}

\item{exprs_tbl_target}{Dataframe with the expression values for the targets in the receiver cell types}

\item{lfc_cutoff}{Cutoff used on the logFC value}

\item{plot_legend}{TRUE (default): add legend to the plot. FALSE: do not add legend.}

\item{heights}{automatic determination if default NULL. If not NULL: number given by the user to indicate the requested heights, which are the height proportions of the different row panels in the plot.}

\item{widths}{automatic determination if default NULL. If not NULL: number given by the user to indicate the requested widths, which are the width proportions of the different columns (side-by-side heatmaps) in the plot.}
}
\value{
List containing the combined heatmap and the legend
}
\description{
\code{make_ligand_activity_target_exprs_plot} Plot the ligand expression in senders plus their activity and target genes in receivers
}
\examples{
\dontrun{
make_ligand_activity_target_exprs_plot(receiver_oi, prioritized_tbl_oi, prioritization_tbl_ligand_receptor, prioritization_tbl_ligand_target, exprs_tbl_ligand, exprs_tbl_target, lfc_cutoff, plot_legend = TRUE, heights = NULL, widths = NULL)
}

}
