% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_nichenet_plotting.R
\name{make_ligand_receptor_lfc_spatial_plot}
\alias{make_ligand_receptor_lfc_spatial_plot}
\title{make_ligand_receptor_lfc_spatial_plot}
\usage{
make_ligand_receptor_lfc_spatial_plot(receiver_oi, prioritized_tbl_oi, prioritization_tbl_ligand_receptor, ligand_spatial = TRUE, receptor_spatial = TRUE, plot_legend = TRUE, heights = NULL, widths = NULL)
}
\arguments{
\item{receiver_oi}{Name of the receiver cell type of interest}

\item{prioritized_tbl_oi}{Dataframe with the ligand-receptor interactions that should be visualized}

\item{prioritization_tbl_ligand_receptor}{$prioritization_tbl_ligand_receptor slot of `get_prioritization_tables`}

\item{ligand_spatial}{TRUE if need to plot the ligand spatial DE info, FALSE if not.}

\item{receptor_spatial}{TRUE if need to plot the receptor spatial DE info, FALSE if not.}

\item{plot_legend}{TRUE (default): add legend to the plot. FALSE: do not add legend.}

\item{heights}{automatic determination if default NULL. If not NULL: number given by the user to indicate the requested heights, which are the height proportions of the different row panels in the plot.}

\item{widths}{automatic determination if default NULL. If not NULL: number given by the user to indicate the requested widths, which are the width proportions of the different columns (side-by-side heatmaps) in the plot.}
}
\value{
List containing the combined heatmap and the legend
}
\description{
\code{make_ligand_receptor_lfc_spatial_plot} Plot the ligand logFC in senders plus the logFC of their receptors in the receivers. In addition, add the spatialDE logFC values of the ligands!
}
\examples{
\dontrun{
make_ligand_receptor_lfc_spatial_plot(receiver_oi, prioritized_tbl_oi, prioritization_tbl_ligand_receptor, ligand_spatial = TRUE, receptor_spatial = TRUE, plot_legend = TRUE, heights = NULL, widths = NULL)
}

}
