% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_visualization.R
\name{make_line_plot}
\alias{make_line_plot}
\title{Make a line plot}
\usage{
make_line_plot(ligand_activities, potential_ligands, ranking_range = c(1, 20))
}
\arguments{
\item{ligand_activities}{Dataframe containing the ligand activities from the sender-agnostic approach}

\item{potential_ligands}{Character vector containing the ligands that are expressed in the sender cell type (i.e. the ligands that are used in the sender-focused approach)}

\item{ranking_range}{Numeric vector of length 2 indicating the range of the rankings to be displayed (default: c(1, 20))}

\item{agnostic_color}{Color representing ligands only inthe sender-agnostic approach (default: "tomato")}

\item{focused_color}{Color representing expressed ligands from the sender-focused approach (default: "black")}

\item{tied_color}{Color to shade ligands that are tied in the same rank (default: "gray75")}

\item{inset_scale}{Numeric value indicating the size of the points and text in the inset (default: 1)}
}
\value{
A ggplot object showing the distribution of sender-focused ligands, as well as a line plot inset comparing the rankings between the two approaches
}
\description{
Make a line plot comparing the ranking of ligands based on their activities in the sender-agnostic and sender-focused approaches
}
\examples{
\dontrun{
# Default
make_line_plot(ligand_activities, potential_ligands)
}
}
