% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{model_evaluation_optimization_nsga2r}
\alias{model_evaluation_optimization_nsga2r}
\title{Construct and evaluate a ligand-target model with the purpose of parameter optimization with NSGA-II.}
\usage{
model_evaluation_optimization_nsga2(y, source_names, algorithm, correct_topology, lr_network, sig_network, gr_network, settings, secondary_targets = FALSE, remove_direct_links = "no",damping_factor = NULL)
}
\arguments{
\item{y}{A numeric vector containing the data source weights as the first elements, and hyperparameters as the last elements. The order of the data source weights accords to the order of source_names.}

\item{source_names}{Character vector containing the names of the data sources. The order of data source names accords to the order of weights in x$source_weights.}

\item{algorithm}{Selection of the algorithm to calculate ligand-tf signaling probability scores. Different options: "PPR" (personalized pagerank), "SPL" (shortest path length) and "direct"(just take weights of ligand-signaling network as ligand-tf weights). Default and recommended: PPR.}

\item{correct_topology}{This parameter indicates whether the PPR-constructed ligand-target matrix will be subtracted by a PR-constructed target matrix. TRUE or FALSE.}

\item{lr_network}{A data frame / tibble containing ligand-receptor interactions (required columns: from, to, source)}

\item{sig_network}{A data frame / tibble containing signaling interactions (required columns: from, to, source)}

\item{gr_network}{A data frame / tibble containing gene regulatory interactions (required columns: from, to, source)}

\item{settings}{A list of lists for which each sub-list contains the following elements: .$name: name of the setting; .$from: name(s) of the ligand(s) active in the setting of interest; .$response: named logical vector indicating whether a target is a TRUE target of the possibly active ligand(s) or a FALSE.}

\item{secondary_targets}{Indicate whether a ligand-target matrix should be returned that explicitly includes putative secondary targets of a ligand (by means of an additional matrix multiplication step considering primary targets as possible regulators). Default: FALSE}

\item{remove_direct_links}{Indicate whether direct ligand-target and receptor-target links in the gene regulatory network should be kept or not. "no": keep links; "ligand": remove direct ligand-target links; "ligand-receptor": remove both direct ligand-target and receptor-target links. Default: "no"}

\item{damping_factor}{The value of the damping factor if damping factor is a fixed parameter and will not be optimized and thus not belong to x. Default NULL.}
}
\value{
A numeric vector of length 4 containing the average auroc for target gene prediction, average aupr (corrected for TP fraction) for target gene prediction, average auroc for ligand activity prediction and average aupr for ligand activity prediction.
}
\description{
\code{model_evaluation_optimization_nsga2} will take as input a vector of data source weights and hyperparameters to construct a ligand-target matrix and evaluate its performance on input validation settings.
}
\examples{
\dontrun{
nr_datasources = source_weights_df$source \%>\% unique() \%>\% length()
test_input = c(rep(0.5, times = nr_datasources), 0.5, 0.5, 0.5, 0.5)
expression_settings_validation = readRDS(url("https://zenodo.org/record/3260758/files/expression_settings.rds"))
test_evaluation_optimization = model_evaluation_optimization_nsga2(test_input, source_weights_df$source \%>\% unique(), "PPR", TRUE, lr_network, sig_network, gr_network,
 lapply(expression_settings_validation, convert_expression_settings_evaluation), secondary_targets = FALSE, remove_direct_links = "no")
}

}
