% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Random sample of something}
\format{A data frame dink
\describe{
  \item{mpg}{price, in US dollars}
  \item{cyl}{weight of the diamond}
  \item{disp}{price, in US dollars}
  \item{hp}{weight of the diamond}
  \item{drat}{price, in US dollars}
  \item{wt}{weight of the diamond}
  \item{qsec}{price, in US dollars}
  \item{vs}{weight of the diamond}
  \item{am}{weight of the diamond}
  \item{gear}{price, in US dollars}
  \item{carb}{weight of the diamond}
  }}
\source{
\url{superdeporte.es}
}
\usage{
mtcars
}
\description{
A dataset containing random samples of something
}
\keyword{datasets}
