% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_nichenet.R
\name{mutate_cond}
\alias{mutate_cond}
\title{Change values in a tibble if some condition is fulfilled.}
\usage{
mutate_cond(.data, condition, ..., envir = parent.frame())
}
\arguments{
\item{.data}{Data frame / tibble}

\item{condition}{A condition that need to be fulfilled.}

\item{...}{The change that need to happen if condition fulfilled -- through the use of `dplyr::mutate()`}

\item{envir}{parent.frame() by default}
}
\value{
A tibble
}
\description{
\code{mutate_cond} Change values in a tibble if some condition is fulfilled. Credits: https://stackoverflow.com/questions/34096162/dplyr-mutate-replace-several-columns-on-a-subset-of-rows.
}
\examples{
\dontrun{
mutate_cond(df, a == 3, b = 4)
}

}
