% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ncitations}
\alias{ncitations}
\title{Number of citations for genes}
\format{
A data frame/tibble
\describe{
  \item{entrez}{entrez}
  \item{ncitations}{ncitations}
  \item{symbol}{symbol}
  \item{entrez_mouse}{entrez_mouse}
  \item{symbol_mouse}{symbol_mouse}
  }
}
\usage{
ncitations
}
\description{
A data.frame/tibble describing the number of times a particular gene is cited in the pubmed literature (date: 13-03-2018)
}
\keyword{datasets}
