% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nichenetr-package.R
\docType{package}
\name{nichenetr-package}
\alias{nichenetr}
\alias{nichenetr-package}
\title{nichenetr: Modeling Intercellular Communication by Linking Ligands to Target Genes with NicheNet}
\description{
This package allows you the investigate intercellular communication from a computational perspective. More specifically, it allows to investigate how interacting cells influence each other's gene expression. Functionalities of this package (e.g. including predicting extracellular upstream regulators and their affected target genes) build upon a probabilistic model of ligand-target links that was inferred by data-integration.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/saeyslab/nichenetr}
  \item Report bugs at \url{https://github.com/saeyslab/nichenetr/issues}
}

}
\author{
\strong{Maintainer}: Chananchida Sang-aram \email{chananchida.sangaram@ugent.be}

Authors:
\itemize{
  \item Robin Browaeys
}

}
\keyword{internal}
