% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{optimized_source_weights_df}
\alias{optimized_source_weights_df}
\title{Optimized data source weights}
\format{
A data frame/tibble
\describe{
  \item{source}{name of data source}
  \item{avg_weight}{weight associated to a data source, averaged over cross-validation folds}
  \item{median_weight}{weight associated to a data source, averaged over cross-validation folds}
  }
}
\usage{
optimized_source_weights_df
}
\description{
A data.frame/tibble describing weight associated to each data source. These weights will be used for weighted aggregation of source networks and were determined via parameter optimization (see NicheNet paper).
}
\keyword{datasets}
