% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{predict_ligand_activities}
\alias{predict_ligand_activities}
\title{Predict activities of ligands in regulating expression of a gene set of interest}
\usage{
predict_ligand_activities(geneset, background_expressed_genes,ligand_target_matrix, potential_ligands, single = TRUE,...)
}
\arguments{
\item{geneset}{Character vector of the gene symbols of genes of which the expression is potentially affected by ligands from the interacting cell.}

\item{background_expressed_genes}{Character vector of gene symbols of the background, non-affected, genes (can contain the symbols of the affected genes as well).}

\item{ligand_target_matrix}{The NicheNet ligand-target matrix denoting regulatory potential scores between ligands and targets (ligands in columns).}

\item{potential_ligands}{Character vector giving the gene symbols of the potentially active ligands you want to define ligand activities for.}

\item{single}{TRUE if you want to calculate ligand activity scores by considering every ligand individually (recommended). FALSE if you want to calculate ligand activity scores as variable importances of a multi-ligand classification model.}

\item{...}{Additional parameters for get_multi_ligand_importances if single = FALSE.}
}
\value{
A tibble giving several ligand activity scores. Following columns in the tibble: $test_ligand, $auroc, $aupr and $pearson.
}
\description{
\code{predict_ligand_activities} Predict activities of ligands in regulating expression of a gene set of interest. Ligand activities are defined as how well they predict the observed transcriptional response (i.e. gene set) according to the NicheNet model.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network,source_weights_df)
ligands = list("TNF","BMP2","IL4")
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands, ltf_cutoff = 0, algorithm = "PPR", damping_factor = 0.5, secondary_targets = FALSE)
potential_ligands = c("TNF","BMP2","IL4")
geneset = c("SOCS2","SOCS3", "IRF1")
background_expressed_genes = c("SOCS2","SOCS3","IRF1","ICAM1","ID1","ID2","ID3")
ligand_activities = predict_ligand_activities(geneset = geneset, background_expressed_genes = background_expressed_genes, ligand_target_matrix = ligand_target_matrix, potential_ligands = potential_ligands)
}

}
