% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_visualization.R
\name{prepare_circos_visualization}
\alias{prepare_circos_visualization}
\title{Prepare circos visualization}
\usage{
prepare_circos_visualization(circos_links, ligand_colors = NULL, target_colors = NULL, widths = NULL, celltype_order = NULL)
}
\arguments{
\item{circos_links}{Dataframe from the function \code{\link{get_ligand_target_links_oi}} containing weighted ligand-target links, cell type expressing the ligand, and target gene goruping}

\item{ligand_colors}{Named vector of colors for each cell type (default: NULL, where colors follow the ggplot default color scheme)}

\item{target_colors}{Named vector of colors for each target gene grouping (default: NULL, where colors follow the ggplot default color scheme)}

\item{widths}{Named list of widths for the different types groupings, including:
\itemize{
\item width_same_cell_same_ligand_type: Width of the links between ligands of the same cell type (default: 0.5)
\item width_different_cell: Width of the links between different cell types, or between different target gene groups (default: 6)
\item width_ligand_target: Width of the links between ligands and targets (default: 15)
\item width_same_cell_same_target_type: Width of the links between target genes of the same group (default: 0.5)
}}

\item{celltype_order}{Order of the cell types (default: NULL, where cell types are ordered alphabetically, followed by "General"). Cell types are then drawn counter-clockwise in the circos plot.}
}
\value{
A list of four objects, including:
\itemize{
\item circos_links: Dataframe of weighted ligand-target links
\item ligand_colors: Named vector of ligands and their colors
\item order: Vector of order of the ligands and target genes
\item gaps: Vector of gaps between the different groupings
}
}
\description{
Prepare the data for the circos visualization by incorporating the colors and order of the links, as well as gaps between different cell types
}
\examples{
\dontrun{
celltype_order <- c("General", "NK", "B", "DC", "Mono")
ligand_colors <- c("General" = "lawngreen", "NK" = "royalblue", "B" = "darkgreen", "Mono" = "violet", "DC" = "steelblue2")
target_colors <- c("LCMV-DE" = "tomato")
vis_circos_obj <- prepare_circos_visualization(circos_links, ligand_colors, target_colors, celltype_order = celltype_order)
}

}
