% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{prepare_ligand_receptor_visualization}
\alias{prepare_ligand_receptor_visualization}
\title{Prepare ligand-receptor visualization}
\usage{
prepare_ligand_receptor_visualization(
  lr_network_top_df_long,
  best_upstream_ligands,
  order_hclust = "both"
)
}
\arguments{
\item{lr_network_top_df_long}{A data frame with three columns, \code{from}, \code{to} and \code{weight}, containing the ligand-receptor interactions and their weights.}

\item{best_upstream_ligands}{Character vector of ligands of interest. This will only be used if \code{order_hclust = "receptors"} or \code{order_hclust = "none"}.}

\item{order_hclust}{"both", "ligands", "receptors", or "none". If "both", the ligands and receptors are ordered by hierarchical clustering. If "ligands" or "receptors" only the ligands or receptors are ordered hierarchically. If "none", no hierarchical clustering is performed, and the ligands are ordered based on \code{best_upstream_ligands}, and the receptors are ordered alphabetically.}
}
\value{
A matrix of ligand-receptor interactions for visualization.
}
\description{
\code{prepare_ligand_receptor_visualization} Prepare a matrix of ligand-receptor interactions for visualization.
}
\examples{
\dontrun{
ligand_receptor_network = prepare_ligand_receptor_visualization(best_upstream_ligands = best_upstream_ligands, expressed_receptors = expressed_receptors, lr_network = lr_network, weighted_networks_lr_sig = weighted_networks_lr_sig, order_hclust = TRUE)
}

}
