% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization_data_sources.R
\name{process_characterization_target_prediction_average}
\alias{process_characterization_target_prediction_average}
\title{Process the output of model evaluation for data source characterization purposes on the target prediction performance (average)}
\usage{
process_characterization_target_prediction_average(output_characterization)
}
\arguments{
\item{output_characterization}{a list of lists containing the results of evaluation of different models (e.g. after execution of \code{evaluate_model}. Every sublist should contain at least the following elements: $model_name and $performances_target_prediction.}
}
\value{
A data frame containing the target gene prediction performances averaged over all validation datasets for all the models that were evaluated.
}
\description{
\code{output_characterization} will process output formed by model evaluation to get a data frame containing performance measures in target gene prediction (averaged over all validation datasets).
}
\examples{

\dontrun{
library(dplyr)
settings = lapply(expression_settings_validation, convert_expression_settings_evaluation)
weights_settings_loi = prepare_settings_leave_one_in_characterization(lr_network,sig_network, gr_network, source_weights_df)
weights_settings_loi = lapply(weights_settings_loi,add_hyperparameters_parameter_settings, lr_sig_hub = 0.25,gr_hub = 0.5,ltf_cutoff = 0,algorithm = "PPR",damping_factor = 0.8,correct_topology = TRUE)
doMC::registerDoMC(cores = 8)
output_characterization = parallel::mclapply(weights_settings_loi[1:3],evaluate_model,lr_network,sig_network, gr_network,settings,calculate_popularity_bias_target_prediction = TRUE, calculate_popularity_bias_ligand_prediction = TRUE, ncitations, mc.cores = 3)
target_prediction_performances = process_characterization_target_prediction_average(output_characterization)
}

}
