% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_nichenet.R
\name{process_niche_de}
\alias{process_niche_de}
\title{Process the DE output of `calculate_niche_de`}
\usage{
process_niche_de(DE_table, niches, type, expression_pct)
}
\arguments{
\item{DE_table}{Output of `calculate_niche_de`}

\item{niches}{a list of lists/niches giving the name, senders and receiver celltypes for each nice. Sender and receiver cell types should be part of Idents(seurat_obj).}

\item{type}{For what type of cellype is the DE analysis: "sender" or "receiver"?}

\item{expression_pct}{Percentage of cells of a cell type having a non-zero expression value for a gene such that a gene can be considered expressed by that cell type.}
}
\value{
A tibble containing processed DE information
}
\description{
\code{process_niche_de} Process the DE output of `calculate_niche_de`: define what the average and minimum logFC value is after comparing a celltype vs all the celltypes of the other niches.
}
\examples{
\dontrun{
seurat_obj = readRDS(url("https://zenodo.org/record/5840787/files/seurat_obj_subset_integrated_zonation.rds"))
niches = list(
"KC_niche" = list(
  "sender" = c("LSECs_portal","Hepatocytes_portal","Stellate cells_portal"),
  "receiver" = c("KCs")),
"MoMac2_niche" = list(
  "sender" = c("Cholangiocytes","Fibroblast 2"),
  "receiver" = c("MoMac2")),
"MoMac1_niche" = list(
  "sender" = c("Capsule fibroblasts","Mesothelial cells"),
 "receiver" = c("MoMac1"))
)
DE_table = calculate_niche_de(seurat_obj, niches, "sender")
process_niche_de(DE_table, niches, "sender",expression_pct = 0.10)
}

}
