% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_random_networks.R
\name{randomize_complete_network_source_specific}
\alias{randomize_complete_network_source_specific}
\title{Randomize an integrated network by shuffling its source networks}
\usage{
randomize_complete_network_source_specific(network)
}
\arguments{
\item{network}{A data frame / tibble containing gene-gene interactions (required columns: $from, $to, $source)}
}
\value{
A randomized network ($from, $to, $source).
}
\description{
\code{randomize_complete_network_source_specific} Randomizes an integrated network of interest by edge swapping in the source-specific networks to preserve the degree of the nodes
}
\examples{
\dontrun{
random_lr = randomize_complete_network_source_specific(lr_network)
}

}
