% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_random_networks.R
\name{randomize_datasource_network}
\alias{randomize_datasource_network}
\title{Randomize a network of a particular data source.}
\usage{
randomize_datasource_network(datasource, network)
}
\arguments{
\item{datasource}{The name of the data source for which the interactions need to be shuffled.}

\item{network}{A data frame / tibble containing gene-gene interactions (required columns: $from, $to, $source)}
}
\value{
A randomized network ($from, $to, $source)
}
\description{
\code{randomize_datasource_network} Randomizes a network of a data source of interest by edge swapping to preserve the degree of the nodes.
}
\examples{
\dontrun{
datasource_lr = lr_network$source[1]
lr_randomized_source = randomize_datasource_network(datasource_lr, lr_network)
}

}
