% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{run_nsga2R_cluster}
\alias{run_nsga2R_cluster}
\title{Run NSGA-II for parameter optimization.}
\usage{
run_nsga2R_cluster(run_id, fn, varNo, objDim, lowerBounds = rep(-Inf, varNo), upperBounds = rep(Inf, varNo), popSize = 100, tourSize = 2, generations = 20, cprob = 0.7, XoverDistIdx = 5, mprob = 0.2, MuDistIdx = 10, ncores = 24)
}
\arguments{
\item{fn}{The function to be optimized, usually \code{model_evaluation_optimization_nsga2}.}

\item{varNo}{The number of variables to be optimized, usually the number of data sources + 4 hyperparameters (lr_sig_hub, gr_hub, ltf_cutoff, damping_factor).}

\item{objDim}{Number of objective functions}

\item{lowerBounds}{A numeric vector containing the lower bounds for the variables to be optimized (default: -Inf)}

\item{upperBounds}{A numeric vector containing the upper bounds for the variables to be optimized (default: Inf)}

\item{popSize}{The population size (default: 100)}

\item{tourSize}{The tournament size (default: 2)}

\item{generations}{The number of generations (default: 20)}

\item{ncores}{The number of cores to be used for parallelization (default: 24)}

\item{cprob}{The crossover probability (default: 0.7)}

\item{XoverDistIdx}{The crossover distribution index, it can be any nonnegative real number (default: 5)}

\item{mprob}{The mutation probability (default: 0.2)}

\item{MuDistIdx}{The mutation distribution index, it can be any nonnegative real number (default: 10)}

\item{...}{Additional arguments to \code{fn}.}
}
\value{
An 'nsga2R' object containing input settings and the following elements:
\itemize{
\item intermed_output_list_params: a list with intermediate values of parameters for each generation (each element has dimensions popSize x varNo)
\item intermed_output_list_obj: a list with intermediate values of objective functions for each generation (each element has dimensions popSize x objDim)
\item parameters: the solutions of variables that were optimized
\item objectives: non-dominated objective function values
\item paretoFrontRank: nondomination ranks (or levels) that each non-dominated solution belongs to
\item crowdingDistance: crowding distances of each non-dominated solution
}
}
\description{
\code{run_nsga2R_cluster} runs the NSGA-II algorithm for parameter optimization and allows for parallelization. The core of this function is adapted from \code{nsga2R::nsga2R}.
}
\examples{
\dontrun{
source_names <- c(lr_network$source, sig_network$source, gr_network$source) \%>\% unique()
n_param <- length(source_names) + 4
n_obj <- 4
lower_bounds <- c(rep(0,times = length(source_names)), 0, 0, 0.9, 0.01)
upper_bounds <- c(rep(1,times = length(source_names)), 1, 1, 0.999, 0.99)
results <- run_nsga2R_cluster(model_evaluation_optimization_nsga2r, varNo=n_param, objDim=n_obj,
lowerBounds=lower_bounds, upperBounds=upper_bounds, popSize = 360, tourSize = 2, generations = 15, ncores = 8)
}

}
