% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_prediction.R
\name{scale_quantile}
\alias{scale_quantile}
\title{Cut off outer quantiles and rescale to a [0, 1] range}
\usage{
scale_quantile(x, outlier_cutoff = .05)
}
\arguments{
\item{x}{A numeric vector, matrix or data frame.}

\item{outlier_cutoff}{The quantile cutoff for outliers (default 0.05).}
}
\value{
The centered, scaled matrix or vector. The numeric centering and scalings used are returned as attributes.
}
\description{
\code{scale_quantile} Cut off outer quantiles and rescale to a [0, 1] range
}
\examples{
\dontrun{
## Generate a matrix from a normal distribution
## with a large standard deviation, centered at c(5, 5)
x <- matrix(rnorm(200*2, sd = 10, mean = 5), ncol = 2)

## Scale the dataset between [0,1]
x_scaled <- scale_quantile(x)

## Show ranges of each column
apply(x_scaled, 2, range)
}
}
