% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differential_nichenet.R
\name{scale_quantile_adapted}
\alias{scale_quantile_adapted}
\title{Normalize values in a vector by quantile scaling and add a pseudovalue of 0.001}
\usage{
scale_quantile_adapted(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A quantile-scaled numeric vector.
}
\description{
\code{scale_quantile_adapted} Normalize values in a vector by quantile scaling. Add a pseudovalue of 0.001 to avoid having a score of 0 for the lowest value.
}
\examples{
\dontrun{
scale_quantile_adapted(rnorm(5))
}

}
