% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaling_modified_zscore}
\alias{scaling_modified_zscore}
\title{Normalize values in a vector by the modified z-score method.}
\usage{
scaling_modified_zscore(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
\code{scaling_modified_zscore} Normalize values by the modified z-score method (uses median and median absolute deviation instead meean)
}
\examples{
scaling_modified_zscore(rnorm(5))

}
