% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scaling_zscore}
\alias{scaling_zscore}
\title{Normalize values in a vector by the z-score method}
\usage{
scaling_zscore(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector.
}
\description{
\code{scaling_zscore} Normalize values in a vector by the z-score method.
}
\examples{
scaling_zscore(rnorm(5))

}
