% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig_network}
\alias{sig_network}
\title{Signaling network}
\format{A data frame/tibble
\describe{
  \item{from}{source gene}
  \item{to}{target gene}
  \item{source}{name of data source describing the interaction}
  }}
\usage{
sig_network
}
\description{
A data.frame/tibble describing gene-gene interactions such as common PPI, phosphorylation,...
}
\keyword{datasets}
