% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{source_weights_df}
\alias{source_weights_df}
\title{Data source weights}
\format{
A data frame/tibble
\describe{
  \item{source}{name of data source}
  \item{weight}{weight associated to a data source}
  }
}
\usage{
source_weights_df
}
\description{
A data.frame/tibble describing weight associated to each data source. These weights will be used for weighted aggregagion of source networks
}
\keyword{datasets}
