% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_optimization.R
\name{visualize_parameter_values}
\alias{visualize_parameter_values}
\title{Visualize parameter values from the output of \code{run_nsga2R_cluster}.}
\usage{
visualize_parameter_values(result_nsga2r, source_names, top_ns = c(5, 10, -10, -25))
}
\arguments{
\item{result_nsga2r}{The output of \code{run_nsga2R_cluster}.}

\item{source_names}{Character vector containing the names of the data sources.}

\item{top_ns}{Numeric vector indicating how many of the best and worst solutions should be considered (negative values indicate the worst solutions; default: c(5, 10, -10, -25)).}
}
\value{
A list containing two ggplot objects, one for the data source weights and one for the hyperparameters.
}
\description{
\code{visualize_parameter_values} will take as input the output of \code{run_nsga2R_cluster} and visualize the data source weights and hyperparameters of the best and worst solutions
}
\examples{
\dontrun{
results <- run_nsga2R_cluster(model_evaluation_optimization_nsga2r, varNo=n_param, objDim=n_obj,
     lowerBounds=lower_bounds, upperBounds=upper_bounds, popSize = 360, tourSize = 2, generations = 15, ncores = 8)

# Visualize the best and worst 5 solutions
visualize_parameter_values(results, source_names, top_ns = c(5, -5))
}

}
