% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{voom_lsec}
\alias{voom_lsec}
\title{Limma voom object corresponding to the ExpressionSet object \code{Exprs_lsec}.}
\format{An object of class \code{EList} with 12173 rows and 12 columns.}
\usage{
voom_lsec
}
\description{
Formal class 'EList'
}
\keyword{datasets}
