% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_target_prediction.R
\name{wrapper_average_performances}
\alias{wrapper_average_performances}
\title{Calculate average performance of datasets of a specific ligand.}
\usage{
wrapper_average_performances(ligand_oi,performances, averaging = "median")
}
\arguments{
\item{ligand_oi}{Name of the ligand for which datasets should be averaged.}

\item{performances}{A data frame with performance values, containing at least folowing variables: $setting, $ligand and one or more metrics.}

\item{averaging}{How should performances of datasets of the same ligand be averaged? 'median' or 'mean'.}
}
\value{
A data frame containing classification evaluation measures for the ligand activity state prediction single, individual feature importance measures.
}
\description{
\code{wrapper_average_performances} Calculate average performance of datasets of a specific ligand. Datasets profiling more than one ligand (and thus ligands other than the ligand of interest), will be included as well.
}
\examples{
\dontrun{
weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df)
settings = lapply(expression_settings_validation[1:5],convert_expression_settings_evaluation)
ligands = extract_ligands_from_settings(settings)
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands)
perf1 = lapply(settings,evaluate_target_prediction,ligand_target_matrix)
performances_target_prediction_averaged = ligands \%>\% lapply(wrapper_average_performances, perf1,"median") \%>\% bind_rows() \%>\% drop_na()
}

}
