% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model_ligand_prediction.R
\name{wrapper_evaluate_single_importances_ligand_prediction}
\alias{wrapper_evaluate_single_importances_ligand_prediction}
\title{Evaluation of ligand activity prediction performance of single ligand importance scores: each dataset individually.}
\usage{
wrapper_evaluate_single_importances_ligand_prediction(group,ligand_importances)
}
\arguments{
\item{group}{Name of the dataset (setting) you want to calculate ligand activity performance for.}

\item{ligand_importances}{A data frame containing at least folowing variables: $setting, $test_ligand, $ligand and one or more feature importance scores. $test_ligand denotes the name of a possibly active ligand, $ligand the name of the truely active ligand.}
}
\value{
A data frame containing classification evaluation measures for the ligand activity state prediction single, individual feature importance measures.
}
\description{
\code{wrapper_evaluate_single_importances_ligand_prediction} Evaluate how well a single ligand importance score is able to predict the true activity state of a ligand. For this it is assumed, that ligand importance measures for truely active ligands will be higher than for non-active ligands. Several classification evaluation metrics for the prediction are calculated and variable importance scores can be extracted to rank the different importance measures in order of importance for ligand activity state prediction.
}
\examples{
\dontrun{
settings = lapply(expression_settings_validation[1:5],convert_expression_settings_evaluation)
settings_ligand_pred = convert_settings_ligand_prediction(settings, all_ligands = unlist(extract_ligands_from_settings(settings,combination = FALSE)), validation = TRUE, single = TRUE)

weighted_networks = construct_weighted_networks(lr_network, sig_network, gr_network, source_weights_df)
ligands = extract_ligands_from_settings(settings_ligand_pred,combination = FALSE)
ligand_target_matrix = construct_ligand_target_matrix(weighted_networks, ligands)
ligand_importances = dplyr::bind_rows(lapply(settings_ligand_pred,get_single_ligand_importances,ligand_target_matrix))
evaluation = ligand_importances$setting \%>\% unique() \%>\% lapply(function(x){x}) \%>\% lapply(wrapper_evaluate_single_importances_ligand_prediction,ligand_importances) \%>\% bind_rows() \%>\% inner_join(ligand_importances \%>\% distinct(setting,ligand))
print(head(evaluation))
}
}
