% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{x}
\alias{x}
\title{Random sample of something}
\format{A vector}
\source{
\url{superdeporte.es}
}
\usage{
x
}
\description{
A dataset containing random samples of something
}
\keyword{datasets}
