% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{calculate_de}
\alias{calculate_de}
\title{Calculate differential expression of one cell type versus all other cell types}
\usage{
calculate_de(seurat_obj, celltype_colname, condition_oi = NA, condition_colname = NA, assay_oi = "RNA", ...)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{celltype_colname}{Name of the meta data column that indicates the cell type of a cell}

\item{condition_oi}{If provided, subset seurat_obj so DE is only calculated for cells belonging to condition_oi}

\item{condition_colname}{Name of the meta data column that indicates from which group/condition a cell comes from}

\item{assay_oi}{Which assay need to be used for DE calculation. Default RNA, alternatives: SCT.}

\item{...}{Arguments passed to Seurat::FindAllMarkers(by default: features = NULL, min.pct = 0, logfc.threshold = 0, return.thresh = 1)}
}
\value{
A dataframe containing the DE results
}
\description{
\code{calculate_de} Calculate differential expression of one cell type versus all other cell types using Seurat::FindAllMarkers. If condition_oi is provided, only consider cells from that condition.
}
\examples{
\dontrun{
seurat_obj = readRDS(url("https://zenodo.org/record/3531889/files/seuratObj.rds"))
seurat_obj$celltype <- make.names(seurat_obj$celltype)
# Calculate cell-type specific markers across conditions
calculate_de(seurat_obj, "celltype")
# Calculate LCMV-specific cell-type markers
calculate_de(seurat_obj, "celltype", condition_oi = "LCMV", condition_colname = "aggregate")
}

}
