% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{convert_human_to_mouse_symbols}
\alias{convert_human_to_mouse_symbols}
\title{Convert human gene symbols to their mouse one-to-one orthologs.}
\usage{
convert_human_to_mouse_symbols(symbols)
}
\arguments{
\item{symbols}{A character vector of official human gene symbols}
}
\value{
A character vector of official mouse gene symbols (one-to-one orthologs of the input human gene symbols).
}
\description{
\code{convert_human_to_mouse_symbols} Convert human gene symbols to their mouse one-to-one orthologs.
}
\examples{
library(dplyr)
human_symbols = c("TNF","IFNG")
mouse_symbols = human_symbols \%>\% convert_human_to_mouse_symbols()
}
