% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneinfo_human}
\alias{geneinfo_human}
\title{Gene annotation information}
\format{
A data frame/tibble
\describe{
  \item{symbol}{human gene symbol}
  \item{entrez}{human gene entrez}
  \item{entrez_mouse}{mouse homolog gene entrez}
  \item{symbol_mouse}{mouse homolog gene symbol}
  }
}
\usage{
geneinfo_human
}
\description{
A data.frame/tibble describing HGNC human gene symbols, their entrez ids and the MGI mouse gene symbols and entrez ids of the one-one orthologs as determined via NCBI's homologene db and biomaRt ensembl db.
}
\keyword{datasets}
