% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_network_extraction.R
\name{get_active_signaling_network}
\alias{get_active_signaling_network}
\title{Get active signaling network in a receiver cell.}
\usage{
get_active_signaling_network(expression_receiver, sig_network, expression_cutoff_receiver = 0)
}
\arguments{
\item{expression_receiver}{A named numeric vector of gene expression levels (absolute or logfc) for the receiver cell that receives extracellular signals from the sender cell}

\item{sig_network}{A data frame / tibble containing signaling interactions (required columns: from, to). Can be both unweighted and weighted.}

\item{expression_cutoff_receiver}{The cutoff on expression value for the receiver cell: receptors will be considered active if their expression is higher than the cutoff. Default: 0.}
}
\value{
A data frame containing at least the variables from, to, receiver_expression. In this network, the active signaling interactions in the system of interest are shown.
}
\description{
\code{get_active_signaling_network} Get active signaling network by looking for which signaling-related genes are expressed in the receiver cell. Instead of looking at absolute expression, it is possible as well to extract a signaling network of differentially expressed signaling mediators if a vector of log2 fold change values is used as input.
}
\examples{
\dontrun{
library(dplyr)
expression_vector_receiver = rnorm(n = 10000, mean = 6, sd = 3)
names(expression_vector_receiver) = sample(x = geneinfo_human$symbol,size = 10000,replace = FALSE)
receiver_sig_network = get_active_signaling_network(expression_vector_receiver,sig_network,expression_cutoff_receiver = 4)
}
}
