% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prioritization.R
\name{get_exprs_avg}
\alias{get_exprs_avg}
\title{Calculate average of gene expression per cell type.}
\usage{
get_exprs_avg(seurat_obj, celltype_colname, condition_oi = NA, condition_colname = NA)
}
\arguments{
\item{seurat_obj}{Seurat object}

\item{celltype_colname}{Name of the meta data column that indicates the cell type of a cell}

\item{condition_oi}{If provided, subset seurat_obj so average expression is only calculated for cells belonging to condition_oi}

\item{condition_colname}{Name of the meta data column that indicates from which group/condition a cell comes from}
}
\value{
Data frame with average gene expression per cell type.
}
\description{
\code{get_exprs_avg}  Calculate average of gene expression per cell type. If condition_oi is provided, only consider cells from that condition.
}
\examples{
\dontrun{
seurat_obj = readRDS(url("https://zenodo.org/record/3531889/files/seuratObj.rds"))
seurat_obj$celltype <- make.names(seuratObj$celltype)
# Calculate average expression across conditions
expression_info = get_exprs_avg(seurat_obj, "celltype")
# Calculate LCMV-specific average expression
expression_info = get_exprs_avg(seurat_obj, "celltype", condition_oi = "LCMV", condition_colname = "aggregate")
}

}
