% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{Random sample of something}
\format{A data frame dink
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond}
  ...
  }}
\source{
\url{superdeporte.es}
}
\usage{
mtcars
}
\description{
A dataset containing random samples of something
}
\keyword{datasets}
