% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sig_network}
\alias{sig_network}
\title{Signaling network}
\format{
A data frame/tibble
\describe{
  \item{from}{source gene}
  \item{to}{target gene}
  \item{source}{name of data source describing the interaction}
  \item{database}{name of the more comprehensive database the data source is part of}
  }
}
\usage{
sig_network
}
\description{
A data.frame/tibble describing gene-gene interactions that are involved in signaling transduction (e.g. PPI, kinase-target,...)
}
\keyword{datasets}
